requirements.txt
# YOLOv8 Space Station Object Detection Requirements
# Core dependencies

# YOLOv8 and PyTorch
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Data Science and Visualization
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# Web Application
streamlit>=1.28.0
streamlit-webrtc>=0.47.0
plotly>=5.17.0

# Utilities
tqdm>=4.65.0
PyYAML>=6.0
albumentations>=1.3.0

# Logging and Monitoring
tensorboard>=2.13.0
wandb>=0.15.0

# Optional: For advanced features
# jupyter>=1.0.0
# ipywidgets>=8.0.0 